+++ ESFramework/Foundation/ESActionBlockContainer.m
@@ -26,4 +26,14 @@
     }
 }
 
+- (SEL)action
+{
+    return @selector(invoke:);
+}
+
++ (SEL)action
+{
+    return @selector(invoke:);
+}
+
 @end